| Questions     | Answers
| ------------- | -------------------------------------------------------
| Description?  | Improved both cold and hot query performance by 2-15x
| Type?         |  improvement
| BC breaks?    | no
| Deprecations? |  no
| Fixed ticket? | Fixes PrestaShop/Prestashop#{issue number here}.
| How to test?  | Please indicate how to best verify that this PR is correct.

Changes:

- id_shop filter will be only added if multishop feature is enabled
- product visibility filter is removed from the base query and added to the outer
- category_product table join on the outer query was removed
- removed all unused selects/joins from base query such as id_manufacture, price, on_sale etc. and added them as joins to $filterToTableMapping.

Performance:
There are the execution times for the queries generated by the module. 
Cold means the faceted blocks are not yet cached, and warm means the blocks are cached, only the base query is running.

- category 1:
  cold: 412ms -> 74ms
  warm: 31ms -> 2ms
- category 2:
  cold: 673ms -> 258ms
  warm: 46ms -> 14ms
- category 3:
  cold: 506ms -> 105ms
  warm: 36ms -> 5ms
- category 4:
  cold: 81ms -> 23ms
  warm: 4ms -> 1ms

Test setup is a dev machine with ryzen 7700x, nvme ssd, 32gb ram, I expect better improvements on weaker configs.