@kpodemski 2024-11-28
Hello @chrisai-dev 

Thank you for your pull request. Very interesting results. Could you share more details about the number of products in a tested category, features, etc.?

Also, it would be great to understand better this part:
> removed all unused selects/joins from base query such as id_manufacture, price, on_sale etc. and added them as joins to $filterToTableMapping.

The rest is fully understandable.