@SharakPL 2024-11-28
> id_shop filter will be only added if multishop feature is enabled

I'm not sure that's wise. What if there are still multishop products in the database, but the user decided to disable multishop temporarily? Won't this make the query include the products from all shops?