@chrisai-dev 2024-11-29
> > id_shop filter will be only added if multishop feature is enabled
> 
> I'm not sure that's wise. What if there are still multishop products in the database, but the user decided to disable multishop temporarily? Won't this make the query include the products from all shops?

@SharakPL, Once you have enabled multishop, and uploaded a product, you cannot disable the feature. No need to add this filter condition for a single store.



> Hello @chrisai-dev
> 
> Thank you for your pull request. Very interesting results. Could you share more details about the number of products in a tested category, features, etc.?
> 
> Also, it would be great to understand better this part:
> 
> > removed all unused selects/joins from base query such as id_manufacture, price, on_sale etc. and added them as joins to $filterToTableMapping.
> 
> The rest is fully understandable.

Hi @kpodemski,

So it's been a while and I've deleted the test results, but, the store roughly had 4k products, 1k categories, the ps_category_product table had about 22000 rows. I tested very large categories with about 1k product, mid sized ones with 100ish, and a small one with 10.

https://github.com/PrestaShop/ps_facetedsearch/blob/79d694e1dcda9a7745677b8827188b1b3fed0bce/src/Adapter/MySQL.php#L789-L802

So this is the part where the fields for the base query are declared. The only field that is "needed no matter what", is id_product, the rest is only needed for specific queries. But these fields had no mappings to tables (https://github.com/PrestaShop/ps_facetedsearch/blob/79d694e1dcda9a7745677b8827188b1b3fed0bce/src/Adapter/MySQL.php#L154)
since they were part of the base query, I added them. 

This is the function that computes the joins based on the query
https://github.com/PrestaShop/ps_facetedsearch/blob/79d694e1dcda9a7745677b8827188b1b3fed0bce/src/Adapter/MySQL.php#L619

I only tested on 8.1.x, don't know if this will affect compatibility with  <= 1.7.