@SharakPL 2024-12-11
> > > id_shop filter will be only added if multishop feature is enabled
> > 
> > 
> > I'm not sure that's wise. What if there are still multishop products in the database, but the user decided to disable multishop temporarily? Won't this make the query include the products from all shops?
> 
> @SharakPL, Once you have enabled multishop, and uploaded a product, you cannot disable the feature. No need to add this filter condition for a single store.

Not sure about PS 8 & 9, but you can disable multishop on 1.7 without problems on Preferences main page. All shops after that work just like before. The only difference is you can't manage additional shops, only the default one.

By removing `id_shop` from filters you'll force the default shop to load all the products from all shops.