@kpodemski 2026-08-29
> [!NOTE]
> Semi-automated review: the benchmarks below were run by Claude Code against a real PrestaShop instance, and every number was checked and signed off by me before posting. A human is in the loop and answering — reply here as you normally would.

Hi @chrisai-dev — first of all, I'm sorry this sat for so long. You opened it in October 2024 with a clear diagnosis and real measurements, and it deserved a much faster answer than this. Thank you for the patience.

The good news is that your analysis was right, and the module ended up going exactly where you pointed. The base query really was dragging unused columns and joins around, and that has since been fixed on `dev` by @Hlavtox in #1303, which measured about a **5× improvement** on a large catalogue. Comparing your branch against `dev` today, the three main ideas are already there:

| your PR | `dev` today |
| --- | --- |
| strip the unused selects from the base query | `setSelectFields(['id_product'])` |
| re-add those fields as mappings | `addRequiredInitialPopulationFields()` — adds only what the outer query asks for |
| `&& getOrderField() != 'position'` | `requiresMappedTable()`, the same idea generalised |

So I rebased your branch onto current `dev` and measured what the remaining delta is worth.

### Setup

PrestaShop 9.1.5 · PHP 8.3 · MariaDB 11.4 · 100 000 products, 400 000 combinations, 4 166-product category. Best-of-3, two rounds with the order reversed, using <https://github.com/kpodemski/prestashop-performance-tester>.

### Results

| page | `dev` | this PR | change |
| --- | ---: | ---: | ---: |
| 100 000-product category | 16.52s | 14.88s | **−10%** |
| 4 166-product category | 1.257s | 1.392s | **+11%** |
| category, sorted by price | 1.149s | 1.393s | **+21%** |
| category, sorted by sales | 1.191s | 1.437s | **+21%** |
| control: product page | 0.165s | 0.166s | +0.9% |
| control: `/search` | 0.050s | 0.050s | +0.6% |

Controls stayed flat, and both rounds agreed, so the differences are real. On top of `dev` it now wins only on the pathological 100 k category and costs 10–20 % on the pages a normal shop actually serves.

### Where that comes from

Splitting the change in two and measuring each half separately: the multishop guard on `id_shop` is **free** (+1.5 %, −0.7 %, 0.0 % — noise). Moving the visibility filter to the outer query carries all of it, and on its own is worse than the full PR:

| page | visibility move only | `id_shop` guard only |
| --- | ---: | ---: |
| sorted by price | +25 … +30% | +1.6% |
| sorted by sales | +18 … +24% | −4.4 … −0.7% |
| category | +13 … +15% | −2.7 … 0.0% |

The reason is that the filter does not really move. Same page, same 481 statements:

| | `dev` | visibility moved out |
| --- | ---: | ---: |
| total SQL time | 0.945s | 1.247s |
| rows examined | 2 055 245 | 2 522 015 |
| `product_shop` joined in the inner query | 22 | 22 |
| `product_shop` joined in the **outer** query | 0 | 8 |

The inner join is still there — `id_shop` keeps it — and the outer query now joins `product_shop` as well, on eight facet and count queries. It is an extra join rather than a relocated one.

### Correctness

For what it's worth, the branch is behaviourally clean: 162 comparable cases (2 categories × 8 filter sets × 8 sort orders × out-of-stock-last on/off) showed **zero** differences in product counts, product order and facet counts. Nothing here is a criticism of the change itself — it simply lands on a very different `dev` than the one you wrote it against.

Given that, I'm going to close this as superseded by #1303. Please don't read that as the work being wasted: you identified the right bottleneck a year before it was fixed, and the fix that shipped is the one you proposed. I'd genuinely welcome more PRs from you.

Caveat, as always: one engine and one data shape. If you have a setup where you see this differently, share the `EXPLAIN` and I'll happily re-open and re-measure.